+++ rate-list/src/main/java/com/dvpermyakov/exchangerate/domain/ExchangeRateEntity.kt
@@ -3,5 +3,5 @@ package com.dvpermyakov.exchangerate.domain
 data class ExchangeRateEntity(
     val fromCode: CurrencyCode,
     val toCode: CurrencyCode,
-    val value: Float
+    val value: Double
 )
